Reorganise files - Move DesktopNav to new Nav dir - add framer-motion - add responsive cols for docs flex layout
