--- docs/components/DesktopNav/index.tsx
@@ -1,1 +1,0 @@
-export { DesktopNav } from './DesktopNav';
