+++ docs/components/Nav/Desktop/DesktopNav.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import { Brand } from '../Global/Brand/Brand';
-import { Tagline } from '../Global/Tagline/Tagline';
+import { Brand } from '../../Global/Brand/Brand';
+import { Tagline } from '../../Global/Tagline/Tagline';
 import { NavContainer, ListItem } from './styles';
 import { DesktopNavLink } from './DesktopNavLink';
-import { links } from './constants';
+import { links } from '../constants';
 
 export function DesktopNav() {
   return (
