+++ docs/components/Nav/Mobile/MobileNav.tsx
@@ -1,0 +1,3 @@
+export function MobileNav() {
+  return <div>Mobile</div>;
+}
