+++ docs/components/Nav/index.tsx
@@ -1,0 +1,1 @@
+export { DesktopNav } from './Desktop/DesktopNav';
